fixed app_controller trying to disable viewable and there is no model

+++ app/app_controller.php
@@ -40,7 +40,7 @@
         {
             parent::beforeFilter();
 
-            //$this->Session->write( 'Auth', ClassRegistry::init( 'Core.User' )->find( 'first', array( 'conditions' => array( 'User.id' => 1 ) ) ) );
+            $this->Session->write( 'Auth', ClassRegistry::init( 'Core.User' )->find( 'first', array( 'conditions' => array( 'User.id' => 1 ) ) ) );
 
             $this->__checkUrl();
             $this->__setupLayout();
@@ -51,7 +51,10 @@
 
             if ( isset( $this->params['prefix'] ) && $this->params['prefix'] == 'admin' && !in_array( $this->params['action'], $this->viewableActions ) )
             {
-                $this->{$this->modelClass}->Behaviors->detach( 'Viewable' );
+                if ( isset( $this->{$this->modelClass}->Behaviors ) )
+                {
+                    $this->{$this->modelClass}->Behaviors->detach( 'Viewable' );
+                }
             }
         }
